Allow adjustment of reward rules (#194)
* refactor farm logic

* rename field

* fix test error

* apply comment from github

* rename file

* fix bug

* fix uint64 overflow

* apply comment from github&change field type uint64 to int64

* Specification error description

* fix bug

* refactor code

* apply comment from github

* format code

* format code

+++ modules/farm/abci.go
@@ -10,7 +10,7 @@ import (
 // EndBlocker handles block beginning logic for farm
 func EndBlocker(ctx sdk.Context, k keeper.Keeper) {
 	logger := k.Logger(ctx).With("handler", "endBlocker")
-	k.IteratorExpiredPool(ctx, uint64(ctx.BlockHeight()), func(pool types.FarmPool) {
+	k.IteratorExpiredPool(ctx, ctx.BlockHeight(), func(pool types.FarmPool) {
 		logger.Info(
 			"The farm pool has expired, refund to creator",
 			"poolName", pool.Name,
